+++ rime/src/widgets/mod.rs
@@ -50,6 +50,6 @@ pub use slider::slider;
 pub use stat::stat;
 pub use status_bar::status_bar;
 pub use stepper::stepper;
-pub use tabs::{tabs, Tab, TAB_BAR_HEIGHT};
+pub use tabs::{tabs, Tab, TabBarStyle, TAB_BAR_HEIGHT};
 pub use toggle::toggle;
 pub use tooltip::{tooltip, Position as TooltipPosition};